// src/app/models/filter.model.ts
export enum SortField {
  FirstName = 'firstName',
  LastName = 'lastName',
  DateOfBirth = 'dateOfBirth'
}

export enum SortOrder {
  Asc = 'asc',
  Desc = 'desc'
}

export enum ActiveFilter {
  All = 'all',
  Active = 'active',
  Inactive = 'inactive'
}

export enum AgeFilter {
  All = 'all',
  Under18 = 'under18',
  Over18 = 'over18'
}